requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
chardet==4.0.0
click==8.1.3
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
tomli==2.0.1
typing-extensions==4.2.0
yarl==1.7.2
